IPHONE-226 Adding support for reading schemas that use namespaces
Node is modified to take an additional NamespaceConcept attribute. As the schema tree is walked, named types that have a namespace push it onto a separate stack in the compiler context. When a named type ends, its namespace is popped off the stack.

Symbolic nodes store the "fullname" of the reference and the symbol map is keyed by "fullname". Some minor gymnastics to patch namespaces and names together where necessary.

Output the namespace for named types that have them.